+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased] - 2024-01-20
+## [Unreleased] - 2024-07-04
 
 ### New Features
 - Add side-effect analysis.
@@ -9,12 +9,19 @@
   - Add `Plugin.onPhaseFinish()`.
   - Support specifying multiple method signatures in one `@InvokeHandler` annotation.
   - Add `getInfo()` to call graph edges and pointer flow edges.
-  - Add pointer analysis assertion mechanism.
+  - Add pointer analysis assertion mechanism to ease testing.
   - Add `pascal.taie.analysis.pta.plugin.util.AnalysisModelPlugin`and `IRModelPlugin` to replace original `Model` and `IRModel`, provide more convenient interfaces to support `@InvokeHandler`.
 - Taint analysis
   - Support specifying IndexRef (e.g., `index: "0[*]"` and `index: "0.f"`) in call sources and parameter sources.
   - Support specifying IndexRef in sinks.
   - Support interactive mode, allowing users to modify the taint configuration file and re-run taint analysis without needing to re-run the whole program analysis.
+  - Enhance TFG dumping by adding taint configuration and call site info to Source/Sink node and TaintTransfer edge
+- Class hierarchy analysis (CHA)
+  - Support ignoring call sites that call methods declared in `java.lang.Object`.
+  - Support ignoring call sites whose callees exceed given limit.
+- Signature pattern and matcher
+  - Add `pascal.taie.language.classes.SignatureMatcher` which supports retrieving classes, methods, or fields whose signature match given pattern.
+  - Use signature matcher in taint analysis and `@InvokeHandler` to simplify signature configuration.
 
 ### Breaking Changes
 - API changes